pyproject.toml
[tool.poetry]
name = "llmia"
version = "0.1.0"
description = "Telegram bot for DnD"
authors = ["Jduun <isaevstanislav2003@mail.ru>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.17.0"
python-dotenv = "^1.0.1"
litellm = "^1.57.8"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
